+++ src/libs/utils/constants.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const os = require('os');
+
+const USER_PERFERENCE_FILE = `${os.homedir}/.serverless-tencent/perference.json`;
+
+module.exports = {
+  USER_PERFERENCE_FILE,
+};