+++ src/libs/utils/index.js
@@ -2,8 +2,10 @@
 
 const basic = require('./basic');
 const utils = require('./utils');
+const constants = require('./constants');
 
 module.exports = {
   ...basic,
   ...utils,
+  ...constants,
 };