+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+# [3.19.3](https://github.com/serverless/serverless-tencent/compare/v3.19.2...v3.19.3) (2022-01-20)
+
+- `help` command will not trigger standalone upgrade
+- Improve standalone ugprade animation
+- The user will not be reminded to upgrade again within seven days after the user chooses not to upgrade.
+- Fix standalone upgrade issue on windows
+- Output detail for `version` command with `binary` or `npm`
+
 # [3.19.2](https://github.com/serverless/serverless-tencent/compare/v3.19.1...v3.19.2) (2022-01-20)
 
 - Fix standalone upgrade on windows